+++ 1_Cross_LowToHiRes/cross_lowtohires.cpp
@@ -244,7 +244,7 @@ int main( int argc, char* argv[] )
 
 	TS.FillFromTrakEM2( gArgs.xml_orig, gArgs.zmin, gArgs.zmax );
 
-	fprintf( flog, "Got %d images.\n", TS.vtil.size() );
+	fprintf( flog, "Got %d images.\n", (int)TS.vtil.size() );
 
 	if( !TS.vtil.size() )
 		goto exit;
