+++ 1_MakeIDB/makeidb.cpp
@@ -614,7 +614,7 @@ int main( int argc, char* argv[] )
 	else
 		TS.FillFromTrakEM2( gArgs.infile, gArgs.zmin, gArgs.zmax );
 
-	fprintf( flog, "Got %d images.\n", TS.vtil.size() );
+	fprintf( flog, "Got %d images.\n", (int)TS.vtil.size() );
 
 	if( !TS.vtil.size() )
 		goto exit;
