+++ 1_Scapeops/scapeops.cpp
@@ -681,7 +681,7 @@ int main( int argc, char* argv[] )
 
 	TS.FillFromTrakEM2( gArgs.infile, gArgs.zb, gArgs.za );
 
-	fprintf( flog, "Got %d images.\n", TS.vtil.size() );
+	fprintf( flog, "Got %d images.\n", (int)TS.vtil.size() );
 
 	if( !TS.vtil.size() )
 		goto exit;
